Move statistics classes to separate module
