+++ src/main/java/com/tungstun/statistics/application/statistics/query/GetBarStatistics.java
@@ -1,4 +1,4 @@
-package com.tungstun.barapi.application.statistics.query;
+package com.tungstun.statistics.application.statistics.query;
 
 import java.util.UUID;
