+++ src/main/java/com/tungstun/statistics/domain/statistics/BarStatistics.java
@@ -1,4 +1,4 @@
-package com.tungstun.barapi.application.statistics.model;
+package com.tungstun.statistics.domain.statistics;
 
 import com.tungstun.barapi.domain.bill.Bill;
 import com.tungstun.barapi.domain.bill.OrderProduct;
