+++ src/main/java/com/tungstun/statistics/domain/statistics/GlobalCustomerStatistics.java
@@ -1,9 +1,9 @@
-package com.tungstun.barapi.application.statistics.model;
+package com.tungstun.statistics.domain.statistics;
 
 import com.tungstun.barapi.domain.bill.Bill;
 import com.tungstun.barapi.domain.bill.OrderProduct;
 
-public record CustomerStatistics(
+public record GlobalCustomerStatistics(
         Double totalSpent,
         Double totalNotYetPayed,
         Bill highestPricedBill,
