+++ src/main/java/com/tungstun/statistics/port/web/statistics/StatisticsController.java
@@ -1,19 +1,19 @@
-package com.tungstun.barapi.port.web.statistics;
+package com.tungstun.statistics.port.web.statistics;
 
-import com.tungstun.barapi.application.statistics.StatisticsQueryHandler;
-import com.tungstun.barapi.application.statistics.model.BarStatistics;
-import com.tungstun.barapi.application.statistics.model.CustomerStatistics;
-import com.tungstun.barapi.application.statistics.model.GlobalCustomerStatistics;
-import com.tungstun.barapi.application.statistics.query.GetBarStatistics;
-import com.tungstun.barapi.application.statistics.query.GetCustomerStatistics;
-import com.tungstun.barapi.application.statistics.query.GetGlobalCustomerStatistics;
-import com.tungstun.barapi.port.web.statistics.converter.BarStatisticsConverter;
-import com.tungstun.barapi.port.web.statistics.converter.CustomerStatisticsConverter;
-import com.tungstun.barapi.port.web.statistics.converter.GlobalCustomerStatisticsConverter;
-import com.tungstun.barapi.port.web.statistics.response.BarStatisticsResponse;
-import com.tungstun.barapi.port.web.statistics.response.CustomerStatisticsResponse;
-import com.tungstun.barapi.port.web.statistics.response.GlobalCustomerStatisticsResponse;
 import com.tungstun.security.domain.user.User;
+import com.tungstun.statistics.application.statistics.StatisticsQueryHandler;
+import com.tungstun.statistics.application.statistics.query.GetBarStatistics;
+import com.tungstun.statistics.application.statistics.query.GetCustomerStatistics;
+import com.tungstun.statistics.application.statistics.query.GetGlobalCustomerStatistics;
+import com.tungstun.statistics.domain.statistics.BarStatistics;
+import com.tungstun.statistics.domain.statistics.CustomerStatistics;
+import com.tungstun.statistics.domain.statistics.GlobalCustomerStatistics;
+import com.tungstun.statistics.port.web.statistics.converter.BarStatisticsConverter;
+import com.tungstun.statistics.port.web.statistics.converter.CustomerStatisticsConverter;
+import com.tungstun.statistics.port.web.statistics.converter.GlobalCustomerStatisticsConverter;
+import com.tungstun.statistics.port.web.statistics.response.BarStatisticsResponse;
+import com.tungstun.statistics.port.web.statistics.response.CustomerStatisticsResponse;
+import com.tungstun.statistics.port.web.statistics.response.GlobalCustomerStatisticsResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import org.springframework.http.HttpStatus;
