+++ src/main/java/com/tungstun/statistics/port/web/statistics/response/BarStatisticsResponse.java
@@ -1,4 +1,4 @@
-package com.tungstun.barapi.port.web.statistics.response;
+package com.tungstun.statistics.port.web.statistics.response;
 
 import com.tungstun.barapi.port.web.bill.response.BillSummaryResponse;
 import com.tungstun.barapi.port.web.order.response.OrderProductResponse;
