+++ src/main/java/com/tungstun/statistics/port/web/statistics/response/CustomerStatisticsResponse.java
@@ -1,9 +1,9 @@
-package com.tungstun.barapi.port.web.statistics.response;
+package com.tungstun.statistics.port.web.statistics.response;
 
 import com.tungstun.barapi.port.web.bill.response.BillSummaryResponse;
 import com.tungstun.barapi.port.web.order.response.OrderProductResponse;
 
-public record GlobalCustomerStatisticsResponse(
+public record CustomerStatisticsResponse(
         Double totalSpent,
         Double totalNotYetPayed,
         BillSummaryResponse highestPricedBill,
