+++ src/main/java/com/tungstun/statistics/port/web/statistics/response/GlobalCustomerStatisticsResponse.java
@@ -1,9 +1,9 @@
-package com.tungstun.barapi.port.web.statistics.response;
+package com.tungstun.statistics.port.web.statistics.response;
 
 import com.tungstun.barapi.port.web.bill.response.BillSummaryResponse;
 import com.tungstun.barapi.port.web.order.response.OrderProductResponse;
 
-public record CustomerStatisticsResponse(
+public record GlobalCustomerStatisticsResponse(
         Double totalSpent,
         Double totalNotYetPayed,
         BillSummaryResponse highestPricedBill,
